Got ESP working with Network web page
ESP can't afford to do keep-alive. Not enough connections. So it now does an HTTP 1.0 response, which doesn't support keep-alive. This causes the web page to close the connection after a response. Also, I was doing the REST calls to localhost and they need to come from document.URL.

Responses are still dummy, but now they show up on the web page