Changed RTOS Wifi to support smartconfig
This is the easiest way to provision the ESP. It attempts to start the network. If successful it sends the NetworkStarted event. If a Disconnect comes in, it means connecting to the network failed. It tries 5 times and then starts smartconfig. You go into the ESP Touch app on a phone, type in the wifi password and press confirm. This sends credentials to the ESP which then restarts the network and everything is up