Moved files around to match new structure
Structure is now that libraries always go in components/<library name>. Makefile and sdkconfig are in the root directory. And the component.mk and all ESP specific files are in the component library directory. To use libm8r in another project you make it a peer of that project and include it from there